refactor files for data plugin
related #11
