+++ test/server/data/queries.js
@@ -1,7 +1,7 @@
 const test = require('tape')
 
 const { init, submitReport } =
-  require('../../../src/server/database/queries.js')
+  require('../../../src/server/data/queries.js')
 const { makeMockQuery } = require('./test_helpers.js')
 
 test('build db', t => {
